requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
openai==1.62.0
requests==2.32.3
scikit-learn==1.6.1
tiktoken==0.8.0
python-dotenv==1.0.1